v1.1.0

+++ src/main/kotlin/us/timinc/mc/cobblemon/fossiltweaks/config/MainConfig.kt
@@ -2,5 +2,5 @@ package us.timinc.mc.cobblemon.fossiltweaks.config
 
 class MainConfig {
     val fossilMachineTicks: Int = 14400
-    val fossilShinyRate: Int = 8192
+    val fossilShinyRate: Float = 8192F
 }
